refactor: arrow 아이콘 tsx에서 svg로 수정
